+++ CHANGELOG.md
@@ -17,6 +17,10 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+- `audio`
+  - [#3713](https://github.com/wix/ricos/pull/3713) fix settings ui & force audio tag usage
+
 </details>
   <hr/>